Broke out combination login and put it in Combine.ts
